// src/ModelManagement/Types/ModelManagementInterfaces.ts
export interface TestingModeGroupResponseDTO {
  testingModeGroupName: string | undefined;
  testingMode: string | undefined;
  description: string | undefined;
  guidId: string;
}